requirements.txt
aerich==0.7.2 ; python_version >= "3.12" and python_version < "4.0"
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.12.13 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.12" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.12" and python_version < "4.0"
aioyoomoney-api==0.0.10 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.6.15 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
dictdiffer==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.104.1 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.7.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.26.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.10.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.5.0 ; python_version >= "3.12" and python_version < "4.0"
openai==1.90.0 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.3.2 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.10.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.12" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "4.0"
; removed python-telegram-bot after migration to aiogram
aiogram==3.15.0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
redis==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
tomlkit==0.13.3 ; python_version >= "3.12" and python_version < "4.0"
tortoise-orm==0.20.1 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.14.0 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.24.0.post1 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.20.1 ; python_version >= "3.12" and python_version < "4.0"
